+++ feature-wallet-impl/src/main/java/jp/co/soramitsu/feature_wallet_impl/data/repository/WalletRepositoryImpl.kt
@@ -174,7 +174,7 @@ class WalletRepositoryImpl(
     }
 
     override suspend fun getAccountFreeBalance(accountAddress: String) =
-        substrateSource.getAccountInfo(accountAddress).data.free.toBigDecimal()
+        substrateSource.getAccountInfo(accountAddress).data.free
 
     private fun defineAccountNameForTransaction(
         accountsByAddress: Map<String, WalletAccount>,
